Implement basic interface support
This should allow interface implementations on records and DUs and
possibly also general instance members. It will generate a function
of the structure 'TypeName.MemberName' which is also embedded into
records and union tuples. We then use the trait call mechanism to
dispatch the interface invokation to the right function. This isn't
particularly efficient as a trait call needs to perform string
concatenation to make the final function name from the type + function
names. If we can find another way to reliably determine which type in a
hierarchy the interface was implemented on then we could generate an
appropriate mod call instead.