Add basic FFI/NFI
Add ModCall attribute to define functions that will be dispatched to the
erlang function defined by the attribute parameters and the arity of the
function.

Also added ErlangTerm attribute to be applied to disciminated unions to
provide a means of defining types that do not carry type information and
are thus more compatible with interop scenarios.