Improve modcall and apply translation.
Keeps a map of module local function to detect if a local function is
called so that we can translate this to an `apply`.
For anything else a modcall is assumed. For member calls (static and
non-static) on an FSharp type we generate the function and call such
that the type is assumed to be present in an fsharp module and has a
type name function prefix. For non-FSharp calls (such as System.DateTime.Now)
we treat the class as a module (e.g. 'System.DateTime':'get_Now'())