Basic query expression support.
Prototype support of query expressions. Currently this only supports
`seq`s. The challenge is how this model would support custom providers,
i.e. query translation. One way would be to use the unevaluated `seq`
structure as the translation source. When encountering a fun, e.g. a
predicate function to a where/filter clause this can be translated to
the erlang abstract format using `erlang:fun_info(env)` which could then
be translated to the target format.