Make function generaration more consistent.
Now members on types are generated as individual modules ending with the
class name. This is a more consistent approach across the board.
Downside is that it generates more modules. Upside is consistency and
type tags in records an unions can now be a single atom which is also
the module that any functions are declared in which means trait calls
and interface dispatch will be much faster.