Object programming
Prototypal implementation of object programming.
Constructor function return a type of {Type, BaseTypes, Closure} where
Type is the standard type tag, the base types a list of all super types
and Closure a map of all fields. Each field key has the structure
`ObjectType.FieldName` in order to guarantee uniqueness in case a
subtype uses a field name of the same name as a super type.

Dispatch is done by checking each type in the object data structure to
see if they export the function and then applying. Not fast I know but
stands a reasonable chance of being fairly correct.